+++ db/migrations/000002_create_urls.up.sql
@@ -1,5 +1,5 @@
 create table URLS(
    URLS_ID serial primary key,
    URLS_ORIGINAL_URL text not null,
-   USERS_ID int not null references USERS(USERS_ID)
+   USERS_ID bigint not null references USERS(USERS_ID)
 );
